+++ packages/smarthr-ui/src/components/AppHeader/components/desktop/AppLauncher.tsx
@@ -3,9 +3,8 @@
 import { type FC, type PropsWithChildren, memo, useCallback, useId, useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useLocalize } from '../../../../hooks/useLocalize'
 import { useTheme } from '../../../../hooks/useTheme'
-import { Localizer } from '../../../../intl'
+import { Localizer, useLocalize } from '../../../../intl'
 import { UnstyledButton } from '../../../Button'
 import { Heading } from '../../../Heading'
 import { FaCircleXmarkIcon, FaStarIcon } from '../../../Icon'
