+++ packages/smarthr-ui/src/components/AppHeader/components/mobile/AppLauncher.tsx
@@ -1,8 +1,7 @@
 import { type FC, type PropsWithChildren, memo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useLocalize } from '../../../../hooks/useLocalize'
-import { Localizer } from '../../../../intl'
+import { Localizer, useLocalize } from '../../../../intl'
 import { UnstyledButton } from '../../../Button'
 import { FaCircleXmarkIcon } from '../../../Icon'
 import { SearchInput } from '../../../Input'
