+++ packages/smarthr-ui/src/components/Combobox/MultiCombobox/MultiCombobox.tsx
@@ -20,9 +20,9 @@ import innerText from 'react-innertext'
 import { tv } from 'tailwind-variants'
 
 import { useLatest } from '../../../hooks/useLatest'
-import { useLocalize } from '../../../hooks/useLocalize'
 import { useOuterClick } from '../../../hooks/useOuterClick'
 import { useTheme } from '../../../hooks/useTheme'
+import { useLocalize } from '../../../intl'
 import { genericsForwardRef } from '../../../libs/util'
 import { FaCaretDownIcon } from '../../Icon'
 import { Scroller } from '../../Scroller'
