+++ packages/smarthr-ui/src/intl/index.ts
@@ -1,5 +1,5 @@
 export { IntlProvider, useAvailableLocales } from './IntlProvider'
-export { useIntl } from './useIntl'
+export { useIntl, useLocalize } from './useIntl'
 export { useDateFormat } from './useDateFormat'
 export { Localizer } from './Localizer'
 export { DateFormatter } from './DateFormatter'
