+++ packages/smarthr-ui/src/intl/useIntl.ts
@@ -32,6 +32,8 @@ export type UseIntlReturn = {
   locale: keyof typeof locales
 }
 
+type LocalizeProps = Parameters<UseIntlReturn['localize']>[0]
+
 const isValidLocale = (locale: string): locale is keyof typeof locales => locale in locales
 
 /**
@@ -70,4 +72,25 @@ export const useIntl = (): UseIntlReturn => {
   )
 
   return result
+}
+
+export const useLocalize = <T extends { [key: string]: LocalizeProps }>(
+  props: T,
+): { [K in keyof T]: string } => {
+  const { localize } = useIntl()
+
+  const localized = useMemo(() => {
+    const result = {} as { [K in keyof T]: string }
+
+    for (const i in props) {
+      result[i] = localize(props[i])
+    }
+
+    return result
+    // HINT: 利用方法としてpropsはliteral objectになる
+    // 依存配列からは意図的に排除している
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [localize])
+
+  return localized
 }
